Move state logger to logger class
